fe: add to cart: disable CSRF protection (it is not necessary and can be annoying)
The CSRF validation used to fail when customer had open page for a long time (days) and then clicked on "add to cart" button.
Moreover, "add to cart" action is done using AJAX so the token was not refreshed to a valid one so repeating the action did not help.