Dependency injection: use "symplify/default-autowire" to autowire all services by default
This is done because Symfony 3.2 considers all services as not autowired by default. Without this bundle we would have to add "autowire: true" to every service definition in services.yml.

Symfony 3.3 will support "_default: autowire: true" so we could remove this bundle after upgrading to Symfony 3.3.
See http://symfony.com/blog/new-in-symfony-3-3-optional-class-for-named-services .