FormTimeProvider: change the time comparison to ">=" because some platforms truncate milliseconds from DateTime
Example: If a form gets rendered at 11:00:00.3 and the form is submitted
at 11:00:05.07, the difference is 5.4 seconds.
However, on some platforms DateTime has the fractional part always zero
and therefore the two times have difference of exactly 5 seconds which
is not greater than TimedFormTypeExtension::OPTION_MINIMUM_SECONDS.

This caused an issue in CustomerRegistrationCest acceptance test which
waits exactly 5 seconds and sometimes this was too little and the test
failed.

With the new "greater than or equal" condition the validation can now
pass in edge cases when the form is rendered at 11:00:00.7 and submitted
at 11:00:05.2 which is a time difference of 4.5 seconds.
I consider this a better behavior then before because even 4 seconds
are still a good countermeasure against bots and if you wait exactly
5.0 seconds it does not bother you with a false validation error.